Split WinRT converters into their own file
